docs: convert SimpleLineChart example to TypeScript (.tsx)
issue: #6041
